requirements.txt
streamlit
tensorflow-cpu
Pillow
numpy
PyGithub
h5py
protobuf
